requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
ccxt==4.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.59.1
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
multidict==6.6.4
nltk==3.9.1
numpy==2.3.2
nvidia-nccl-cu12==2.27.7
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
pycares==4.10.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.41
ta==0.11.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==3.0.4
yarl==1.20.1
